Simplified api.validate_id(id) regex.
api.validate_id(id) does not need to return `spotify` or the 22-digit `id` in the group. Running the statid method is for the sole purpose of validating if the ID is a properly formatted Spotify URI and to get the `model` from it for use with `api.lookup_id(cls, id)`.